refactor: extract note builder utilities from useNoteActions — separate pure functions from React hook
